Small fixes and additions (#12)
* Remove optimality @assert's

Even if we don't solve the Lagrangian relaxation to optimality, we still
have a valid multiplier and a valid cut.  So we don't need to fail the
whole SDDiP solution.

This should be warned to the user, though.

* English Typos

* Name binary state variables with _bin_

Uses JuMP names of the original variables, so now we can see them more
clearly in the problem's equations when printed.